Add favorites into a playlist (#10)
* changes for favorite option

* adding features for favorites

* chore: remove unused app_config.dart import from other_screen.dart.

* style: apply consistent stadium borders and theme-based color schemes to buttons across playlist screens and modals

* chore: remove uuid dependency and fix trailing comma in localization file

* feat: add editEnrollmentLabel localization support for English and Marathi

* refactor: simplify MaterialApp configuration by removing redundant route definitions and nesting

---------

+++ lib/jap_mala/widgets/manual_jap_tab.dart
@@ -256,7 +256,7 @@ class _ManualJapTabState extends State<ManualJapTab>
                                 return Positioned(
                                   top: topOffset,
                                   child: Opacity(
-                                    opacity: opacity as double,
+                                    opacity: opacity,
                                     child: SizedBox(
                                       width: 64,
                                       height: 95,
